pyproject.toml
[project]
name = "ai-video-analyze-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "discord-py>=2.5.2",
    "python-dotenv>=1.0.1",
    "openai>=1.35.13",
    "yt-dlp>=2024.7.25",
]
